requirements.txt
Flask==2.0.1
newspaper3k==0.2.8
Werkzeug==2.0.1